// src/enums/AnnotationOptionEnum.ts
export const enum AnnotationOptionEnum {
  // 自由画笔
  INK_1 = "INK1",
  INK_2 = "INK2",
  INK_3 = "INK3",
  INK_4 = "INK4",
  // 文本
  TEXT_1 = "TEXT1",
  TEXT_2 = "TEXT2",
  TEXT_3 = "TEXT3",
  // 文本标记
  TEXTMARK_1 = "TEXTMARK1",
  TEXTMARK_2 = "TEXTMARK2",
  TEXTMARK_3 = "TEXTMARK3",
  TEXTMARK_4 = "TEXTMARK4",
  // 形状
  SHAPE_1 = "SHAPE1",
  SHAPE_2 = "SHAPE2",
  SHAPE_3 = "SHAPE3",
  SHAPE_4 = "SHAPE4",
  // 笔记
  NOTE_1 = "NOTE1",
  NOTE_2 = "NOTE2",
  NOTE_3 = "NOTE3",
  NOTE_4 = "NOTE4",
}